reestructure dictionary information
