+++ modules/aruco/samples/marker_detector_pose.cpp
