+++ modules/aruco/src/predefined_dictionaries.cpp
@@ -1,0 +1,232 @@
+/*
+By downloading, copying, installing or using the software you agree to this
+license. If you do not agree to this license, do not download, install,
+copy or use the software.
+
+                          License Agreement
+               For Open Source Computer Vision Library
+                       (3-clause BSD License)
+
+Copyright (C) 2013, OpenCV Foundation, all rights reserved.
+Third party copyrights are property of their respective owners.
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+  * Redistributions of source code must retain the above copyright notice,
+    this list of conditions and the following disclaimer.
+
+  * Redistributions in binary form must reproduce the above copyright notice,
+    this list of conditions and the following disclaimer in the documentation
+    and/or other materials provided with the distribution.
+
+  * Neither the names of the copyright holders nor the names of the contributors
+    may be used to endorse or promote products derived from this software
+    without specific prior written permission.
+
+This software is provided by the copyright holders and contributors "as is" and
+any express or implied warranties, including, but not limited to, the implied
+warranties of merchantability and fitness for a particular purpose are
+disclaimed. In no event shall copyright holders or contributors be liable for
+any direct, indirect, incidental, special, exemplary, or consequential damages
+(including, but not limited to, procurement of substitute goods or services;
+loss of use, data, or profits; or business interruption) however caused
+and on any theory of liability, whether in contract, strict liability,
+or tort (including negligence or otherwise) arising in any way out of
+the use of this software, even if advised of the possibility of such damage.
+*/
+
+#ifndef __OPENCV_ARUCO_PREDEFINED_DICTIONARIES_CPP__
+#define __OPENCV_ARUCO_PREDEFINED_DICTIONARIES_CPP__
+#ifdef __cplusplus
+
+#include "opencv2/dictionary.hpp"
+
+
+namespace cv{ namespace aruco{
+
+
+const char _dict_aruco_quartets[][7] = {
+    {9,1,1,1,0,0,0}, {11,3,3,1,0,0,0}, {10,1,1,3,0,0,0}, {8,3,3,3,0,0,0}, {9,1,1,5,2,2,0}, {11,3,3,5,2,2,0},
+    {10,1,1,7,2,2,0}, {8,3,3,7,2,2,0}, {9,0,1,5,1,0,0}, {11,2,3,5,1,0,0}, {10,0,1,7,1,0,0}, {8,2,3,7,1,0,0},
+    {9,0,1,1,3,2,0}, {11,2,3,1,3,2,0}, {10,0,1,3,3,2,0}, {8,2,3,3,3,2,0}, {9,1,5,1,0,4,15}, {11,3,7,1,0,4,15},
+    {10,1,5,3,0,4,15}, {8,3,7,3,0,4,15}, {9,1,5,5,2,6,15}, {11,3,7,5,2,6,15}, {10,1,5,7,2,6,15}, {8,3,7,7,2,6,15},
+    {9,0,5,5,1,4,15}, {11,2,7,5,1,4,15}, {10,0,5,7,1,4,15}, {8,2,7,7,1,4,15}, {9,0,5,1,3,6,15}, {11,2,7,1,3,6,15},
+    {10,0,5,3,3,6,15}, {8,2,7,3,3,6,15}, {9,1,4,1,0,1,0}, {11,3,6,1,0,1,0}, {10,1,4,3,0,1,0}, {8,3,6,3,0,1,0},
+    {9,1,4,5,2,3,0}, {11,3,6,5,2,3,0}, {10,1,4,7,2,3,0}, {8,3,6,7,2,3,0}, {9,0,4,5,1,1,0}, {11,2,6,5,1,1,0},
+    {10,0,4,7,1,1,0}, {8,2,6,7,1,1,0}, {9,0,4,1,3,3,0}, {11,2,6,1,3,3,0}, {10,0,4,3,3,3,0}, {8,2,6,3,3,3,0},
+    {9,1,0,1,0,5,15}, {11,3,2,1,0,5,15}, {10,1,0,3,0,5,15}, {8,3,2,3,0,5,15}, {9,1,0,5,2,7,15}, {11,3,2,5,2,7,15},
+    {10,1,0,7,2,7,15}, {8,3,2,7,2,7,15}, {9,0,0,5,1,5,15}, {11,2,2,5,1,5,15}, {10,0,0,7,1,5,15}, {8,2,2,7,1,5,15},
+    {9,0,0,1,3,7,15}, {11,2,2,1,3,7,15}, {10,0,0,3,3,7,15}, {8,2,2,3,3,7,15}, {9,5,1,1,4,8,0}, {11,7,3,1,4,8,0},
+    {10,5,1,3,4,8,0}, {8,7,3,3,4,8,0}, {9,5,1,5,6,10,0}, {11,7,3,5,6,10,0}, {10,5,1,7,6,10,0}, {8,7,3,7,6,10,0},
+    {9,4,1,5,5,8,0}, {11,6,3,5,5,8,0}, {10,4,1,7,5,8,0}, {8,6,3,7,5,8,0}, {9,4,1,1,7,10,0}, {11,6,3,1,7,10,0},
+    {10,4,1,3,7,10,0}, {8,6,3,3,7,10,0}, {9,5,5,1,4,12,15}, {11,7,7,1,4,12,15}, {10,5,5,3,4,12,15}, {8,7,7,3,4,12,15},
+    {9,5,5,5,6,14,15}, {11,7,7,5,6,14,15}, {10,5,5,7,6,14,15}, {8,7,7,7,6,14,15}, {9,4,5,5,5,12,15}, {11,6,7,5,5,12,15},
+    {10,4,5,7,5,12,15}, {8,6,7,7,5,12,15}, {9,4,5,1,7,14,15}, {11,6,7,1,7,14,15}, {10,4,5,3,7,14,15}, {8,6,7,3,7,14,15},
+    {9,5,4,1,4,9,0}, {11,7,6,1,4,9,0}, {10,5,4,3,4,9,0}, {8,7,6,3,4,9,0}, {9,5,4,5,6,11,0}, {11,7,6,5,6,11,0},
+    {10,5,4,7,6,11,0}, {8,7,6,7,6,11,0}, {9,4,4,5,5,9,0}, {11,6,6,5,5,9,0}, {10,4,4,7,5,9,0}, {8,6,6,7,5,9,0},
+    {9,4,4,1,7,11,0}, {11,6,6,1,7,11,0}, {10,4,4,3,7,11,0}, {8,6,6,3,7,11,0}, {9,5,0,1,4,13,15}, {11,7,2,1,4,13,15},
+    {10,5,0,3,4,13,15}, {8,7,2,3,4,13,15}, {9,5,0,5,6,15,15}, {11,7,2,5,6,15,15}, {10,5,0,7,6,15,15}, {8,7,2,7,6,15,15},
+    {9,4,0,5,5,13,15}, {11,6,2,5,5,13,15}, {10,4,0,7,5,13,15}, {8,6,2,7,5,13,15}, {9,4,0,1,7,15,15}, {11,6,2,1,7,15,15},
+    {10,4,0,3,7,15,15}, {8,6,2,3,7,15,15}, {9,5,1,0,8,0,0}, {11,7,3,0,8,0,0}, {10,5,1,2,8,0,0}, {8,7,3,2,8,0,0},
+    {9,5,1,4,10,2,0}, {11,7,3,4,10,2,0}, {10,5,1,6,10,2,0}, {8,7,3,6,10,2,0}, {9,4,1,4,9,0,0}, {11,6,3,4,9,0,0},
+    {10,4,1,6,9,0,0}, {8,6,3,6,9,0,0}, {9,4,1,0,11,2,0}, {11,6,3,0,11,2,0}, {10,4,1,2,11,2,0}, {8,6,3,2,11,2,0},
+    {9,5,5,0,8,4,15}, {11,7,7,0,8,4,15}, {10,5,5,2,8,4,15}, {8,7,7,2,8,4,15}, {9,5,5,4,10,6,15}, {11,7,7,4,10,6,15},
+    {10,5,5,6,10,6,15}, {8,7,7,6,10,6,15}, {9,4,5,4,9,4,15}, {11,6,7,4,9,4,15}, {10,4,5,6,9,4,15}, {8,6,7,6,9,4,15},
+    {9,4,5,0,11,6,15}, {11,6,7,0,11,6,15}, {10,4,5,2,11,6,15}, {8,6,7,2,11,6,15}, {9,5,4,0,8,1,0}, {11,7,6,0,8,1,0},
+    {10,5,4,2,8,1,0}, {8,7,6,2,8,1,0}, {9,5,4,4,10,3,0}, {11,7,6,4,10,3,0}, {10,5,4,6,10,3,0}, {8,7,6,6,10,3,0},
+    {9,4,4,4,9,1,0}, {11,6,6,4,9,1,0}, {10,4,4,6,9,1,0}, {8,6,6,6,9,1,0}, {9,4,4,0,11,3,0}, {11,6,6,0,11,3,0},
+    {10,4,4,2,11,3,0}, {8,6,6,2,11,3,0}, {9,5,0,0,8,5,15}, {11,7,2,0,8,5,15}, {10,5,0,2,8,5,15}, {8,7,2,2,8,5,15},
+    {9,5,0,4,10,7,15}, {11,7,2,4,10,7,15}, {10,5,0,6,10,7,15}, {8,7,2,6,10,7,15}, {9,4,0,4,9,5,15}, {11,6,2,4,9,5,15},
+    {10,4,0,6,9,5,15}, {8,6,2,6,9,5,15}, {9,4,0,0,11,7,15}, {11,6,2,0,11,7,15}, {10,4,0,2,11,7,15}, {8,6,2,2,11,7,15},
+    {9,1,1,0,12,8,0}, {11,3,3,0,12,8,0}, {10,1,1,2,12,8,0}, {8,3,3,2,12,8,0}, {9,1,1,4,14,10,0}, {11,3,3,4,14,10,0},
+    {10,1,1,6,14,10,0}, {8,3,3,6,14,10,0}, {9,0,1,4,13,8,0}, {11,2,3,4,13,8,0}, {10,0,1,6,13,8,0}, {8,2,3,6,13,8,0},
+    {9,0,1,0,15,10,0}, {11,2,3,0,15,10,0}, {10,0,1,2,15,10,0}, {8,2,3,2,15,10,0}, {9,1,5,0,12,12,15}, {11,3,7,0,12,12,15},
+    {10,1,5,2,12,12,15}, {8,3,7,2,12,12,15}, {9,1,5,4,14,14,15}, {11,3,7,4,14,14,15}, {10,1,5,6,14,14,15}, {8,3,7,6,14,14,15},
+    {9,0,5,4,13,12,15}, {11,2,7,4,13,12,15}, {10,0,5,6,13,12,15}, {8,2,7,6,13,12,15}, {9,0,5,0,15,14,15}, {11,2,7,0,15,14,15},
+    {10,0,5,2,15,14,15}, {8,2,7,2,15,14,15}, {9,1,4,0,12,9,0}, {11,3,6,0,12,9,0}, {10,1,4,2,12,9,0}, {8,3,6,2,12,9,0},
+    {9,1,4,4,14,11,0}, {11,3,6,4,14,11,0}, {10,1,4,6,14,11,0}, {8,3,6,6,14,11,0}, {9,0,4,4,13,9,0}, {11,2,6,4,13,9,0},
+    {10,0,4,6,13,9,0}, {8,2,6,6,13,9,0}, {9,0,4,0,15,11,0}, {11,2,6,0,15,11,0}, {10,0,4,2,15,11,0}, {8,2,6,2,15,11,0},
+    {9,1,0,0,12,13,15}, {11,3,2,0,12,13,15}, {10,1,0,2,12,13,15}, {8,3,2,2,12,13,15}, {9,1,0,4,14,15,15}, {11,3,2,4,14,15,15},
+    {10,1,0,6,14,15,15}, {8,3,2,6,14,15,15}, {9,0,0,4,13,13,15}, {11,2,2,4,13,13,15}, {10,0,0,6,13,13,15}, {8,2,2,6,13,13,15},
+    {9,0,0,0,15,15,15}, {11,2,2,0,15,15,15}, {10,0,0,2,15,15,15}, {8,2,2,2,15,15,15}, {13,1,9,9,0,0,0}, {15,3,11,9,0,0,0},
+    {14,1,9,11,0,0,0}, {12,3,11,11,0,0,0}, {13,1,9,13,2,2,0}, {15,3,11,13,2,2,0}, {14,1,9,15,2,2,0}, {12,3,11,15,2,2,0},
+    {13,0,9,13,1,0,0}, {15,2,11,13,1,0,0}, {14,0,9,15,1,0,0}, {12,2,11,15,1,0,0}, {13,0,9,9,3,2,0}, {15,2,11,9,3,2,0},
+    {14,0,9,11,3,2,0}, {12,2,11,11,3,2,0}, {13,1,13,9,0,4,15}, {15,3,15,9,0,4,15}, {14,1,13,11,0,4,15}, {12,3,15,11,0,4,15},
+    {13,1,13,13,2,6,15}, {15,3,15,13,2,6,15}, {14,1,13,15,2,6,15}, {12,3,15,15,2,6,15}, {13,0,13,13,1,4,15}, {15,2,15,13,1,4,15},
+    {14,0,13,15,1,4,15}, {12,2,15,15,1,4,15}, {13,0,13,9,3,6,15}, {15,2,15,9,3,6,15}, {14,0,13,11,3,6,15}, {12,2,15,11,3,6,15},
+    {13,1,12,9,0,1,0}, {15,3,14,9,0,1,0}, {14,1,12,11,0,1,0}, {12,3,14,11,0,1,0}, {13,1,12,13,2,3,0}, {15,3,14,13,2,3,0},
+    {14,1,12,15,2,3,0}, {12,3,14,15,2,3,0}, {13,0,12,13,1,1,0}, {15,2,14,13,1,1,0}, {14,0,12,15,1,1,0}, {12,2,14,15,1,1,0},
+    {13,0,12,9,3,3,0}, {15,2,14,9,3,3,0}, {14,0,12,11,3,3,0}, {12,2,14,11,3,3,0}, {13,1,8,9,0,5,15}, {15,3,10,9,0,5,15},
+    {14,1,8,11,0,5,15}, {12,3,10,11,0,5,15}, {13,1,8,13,2,7,15}, {15,3,10,13,2,7,15}, {14,1,8,15,2,7,15}, {12,3,10,15,2,7,15},
+    {13,0,8,13,1,5,15}, {15,2,10,13,1,5,15}, {14,0,8,15,1,5,15}, {12,2,10,15,1,5,15}, {13,0,8,9,3,7,15}, {15,2,10,9,3,7,15},
+    {14,0,8,11,3,7,15}, {12,2,10,11,3,7,15}, {13,5,9,9,4,8,0}, {15,7,11,9,4,8,0}, {14,5,9,11,4,8,0}, {12,7,11,11,4,8,0},
+    {13,5,9,13,6,10,0}, {15,7,11,13,6,10,0}, {14,5,9,15,6,10,0}, {12,7,11,15,6,10,0}, {13,4,9,13,5,8,0}, {15,6,11,13,5,8,0},
+    {14,4,9,15,5,8,0}, {12,6,11,15,5,8,0}, {13,4,9,9,7,10,0}, {15,6,11,9,7,10,0}, {14,4,9,11,7,10,0}, {12,6,11,11,7,10,0},
+    {13,5,13,9,4,12,15}, {15,7,15,9,4,12,15}, {14,5,13,11,4,12,15}, {12,7,15,11,4,12,15}, {13,5,13,13,6,14,15}, {15,7,15,13,6,14,15},
+    {14,5,13,15,6,14,15}, {12,7,15,15,6,14,15}, {13,4,13,13,5,12,15}, {15,6,15,13,5,12,15}, {14,4,13,15,5,12,15}, {12,6,15,15,5,12,15},
+    {13,4,13,9,7,14,15}, {15,6,15,9,7,14,15}, {14,4,13,11,7,14,15}, {12,6,15,11,7,14,15}, {13,5,12,9,4,9,0}, {15,7,14,9,4,9,0},
+    {14,5,12,11,4,9,0}, {12,7,14,11,4,9,0}, {13,5,12,13,6,11,0}, {15,7,14,13,6,11,0}, {14,5,12,15,6,11,0}, {12,7,14,15,6,11,0},
+    {13,4,12,13,5,9,0}, {15,6,14,13,5,9,0}, {14,4,12,15,5,9,0}, {12,6,14,15,5,9,0}, {13,4,12,9,7,11,0}, {15,6,14,9,7,11,0},
+    {14,4,12,11,7,11,0}, {12,6,14,11,7,11,0}, {13,5,8,9,4,13,15}, {15,7,10,9,4,13,15}, {14,5,8,11,4,13,15}, {12,7,10,11,4,13,15},
+    {13,5,8,13,6,15,15}, {15,7,10,13,6,15,15}, {14,5,8,15,6,15,15}, {12,7,10,15,6,15,15}, {13,4,8,13,5,13,15}, {15,6,10,13,5,13,15},
+    {14,4,8,15,5,13,15}, {12,6,10,15,5,13,15}, {13,4,8,9,7,15,15}, {15,6,10,9,7,15,15}, {14,4,8,11,7,15,15}, {12,6,10,11,7,15,15},
+    {13,5,9,8,8,0,0}, {15,7,11,8,8,0,0}, {14,5,9,10,8,0,0}, {12,7,11,10,8,0,0}, {13,5,9,12,10,2,0}, {15,7,11,12,10,2,0},
+    {14,5,9,14,10,2,0}, {12,7,11,14,10,2,0}, {13,4,9,12,9,0,0}, {15,6,11,12,9,0,0}, {14,4,9,14,9,0,0}, {12,6,11,14,9,0,0},
+    {13,4,9,8,11,2,0}, {15,6,11,8,11,2,0}, {14,4,9,10,11,2,0}, {12,6,11,10,11,2,0}, {13,5,13,8,8,4,15}, {15,7,15,8,8,4,15},
+    {14,5,13,10,8,4,15}, {12,7,15,10,8,4,15}, {13,5,13,12,10,6,15}, {15,7,15,12,10,6,15}, {14,5,13,14,10,6,15}, {12,7,15,14,10,6,15},
+    {13,4,13,12,9,4,15}, {15,6,15,12,9,4,15}, {14,4,13,14,9,4,15}, {12,6,15,14,9,4,15}, {13,4,13,8,11,6,15}, {15,6,15,8,11,6,15},
+    {14,4,13,10,11,6,15}, {12,6,15,10,11,6,15}, {13,5,12,8,8,1,0}, {15,7,14,8,8,1,0}, {14,5,12,10,8,1,0}, {12,7,14,10,8,1,0},
+    {13,5,12,12,10,3,0}, {15,7,14,12,10,3,0}, {14,5,12,14,10,3,0}, {12,7,14,14,10,3,0}, {13,4,12,12,9,1,0}, {15,6,14,12,9,1,0},
+    {14,4,12,14,9,1,0}, {12,6,14,14,9,1,0}, {13,4,12,8,11,3,0}, {15,6,14,8,11,3,0}, {14,4,12,10,11,3,0}, {12,6,14,10,11,3,0},
+    {13,5,8,8,8,5,15}, {15,7,10,8,8,5,15}, {14,5,8,10,8,5,15}, {12,7,10,10,8,5,15}, {13,5,8,12,10,7,15}, {15,7,10,12,10,7,15},
+    {14,5,8,14,10,7,15}, {12,7,10,14,10,7,15}, {13,4,8,12,9,5,15}, {15,6,10,12,9,5,15}, {14,4,8,14,9,5,15}, {12,6,10,14,9,5,15},
+    {13,4,8,8,11,7,15}, {15,6,10,8,11,7,15}, {14,4,8,10,11,7,15}, {12,6,10,10,11,7,15}, {13,1,9,8,12,8,0}, {15,3,11,8,12,8,0},
+    {14,1,9,10,12,8,0}, {12,3,11,10,12,8,0}, {13,1,9,12,14,10,0}, {15,3,11,12,14,10,0}, {14,1,9,14,14,10,0}, {12,3,11,14,14,10,0},
+    {13,0,9,12,13,8,0}, {15,2,11,12,13,8,0}, {14,0,9,14,13,8,0}, {12,2,11,14,13,8,0}, {13,0,9,8,15,10,0}, {15,2,11,8,15,10,0},
+    {14,0,9,10,15,10,0}, {12,2,11,10,15,10,0}, {13,1,13,8,12,12,15}, {15,3,15,8,12,12,15}, {14,1,13,10,12,12,15}, {12,3,15,10,12,12,15},
+    {13,1,13,12,14,14,15}, {15,3,15,12,14,14,15}, {14,1,13,14,14,14,15}, {12,3,15,14,14,14,15}, {13,0,13,12,13,12,15}, {15,2,15,12,13,12,15},
+    {14,0,13,14,13,12,15}, {12,2,15,14,13,12,15}, {13,0,13,8,15,14,15}, {15,2,15,8,15,14,15}, {14,0,13,10,15,14,15}, {12,2,15,10,15,14,15},
+    {13,1,12,8,12,9,0}, {15,3,14,8,12,9,0}, {14,1,12,10,12,9,0}, {12,3,14,10,12,9,0}, {13,1,12,12,14,11,0}, {15,3,14,12,14,11,0},
+    {14,1,12,14,14,11,0}, {12,3,14,14,14,11,0}, {13,0,12,12,13,9,0}, {15,2,14,12,13,9,0}, {14,0,12,14,13,9,0}, {12,2,14,14,13,9,0},
+    {13,0,12,8,15,11,0}, {15,2,14,8,15,11,0}, {14,0,12,10,15,11,0}, {12,2,14,10,15,11,0}, {13,1,8,8,12,13,15}, {15,3,10,8,12,13,15},
+    {14,1,8,10,12,13,15}, {12,3,10,10,12,13,15}, {13,1,8,12,14,15,15}, {15,3,10,12,14,15,15}, {14,1,8,14,14,15,15}, {12,3,10,14,14,15,15},
+    {13,0,8,12,13,13,15}, {15,2,10,12,13,13,15}, {14,0,8,14,13,13,15}, {12,2,10,14,13,13,15}, {13,0,8,8,15,15,15}, {15,2,10,8,15,15,15},
+    {14,0,8,10,15,15,15}, {12,2,10,10,15,15,15}, {5,9,1,1,0,0,0}, {7,11,3,1,0,0,0}, {6,9,1,3,0,0,0}, {4,11,3,3,0,0,0},
+    {5,9,1,5,2,2,0}, {7,11,3,5,2,2,0}, {6,9,1,7,2,2,0}, {4,11,3,7,2,2,0}, {5,8,1,5,1,0,0}, {7,10,3,5,1,0,0},
+    {6,8,1,7,1,0,0}, {4,10,3,7,1,0,0}, {5,8,1,1,3,2,0}, {7,10,3,1,3,2,0}, {6,8,1,3,3,2,0}, {4,10,3,3,3,2,0},
+    {5,9,5,1,0,4,15}, {7,11,7,1,0,4,15}, {6,9,5,3,0,4,15}, {4,11,7,3,0,4,15}, {5,9,5,5,2,6,15}, {7,11,7,5,2,6,15},
+    {6,9,5,7,2,6,15}, {4,11,7,7,2,6,15}, {5,8,5,5,1,4,15}, {7,10,7,5,1,4,15}, {6,8,5,7,1,4,15}, {4,10,7,7,1,4,15},
+    {5,8,5,1,3,6,15}, {7,10,7,1,3,6,15}, {6,8,5,3,3,6,15}, {4,10,7,3,3,6,15}, {5,9,4,1,0,1,0}, {7,11,6,1,0,1,0},
+    {6,9,4,3,0,1,0}, {4,11,6,3,0,1,0}, {5,9,4,5,2,3,0}, {7,11,6,5,2,3,0}, {6,9,4,7,2,3,0}, {4,11,6,7,2,3,0},
+    {5,8,4,5,1,1,0}, {7,10,6,5,1,1,0}, {6,8,4,7,1,1,0}, {4,10,6,7,1,1,0}, {5,8,4,1,3,3,0}, {7,10,6,1,3,3,0},
+    {6,8,4,3,3,3,0}, {4,10,6,3,3,3,0}, {5,9,0,1,0,5,15}, {7,11,2,1,0,5,15}, {6,9,0,3,0,5,15}, {4,11,2,3,0,5,15},
+    {5,9,0,5,2,7,15}, {7,11,2,5,2,7,15}, {6,9,0,7,2,7,15}, {4,11,2,7,2,7,15}, {5,8,0,5,1,5,15}, {7,10,2,5,1,5,15},
+    {6,8,0,7,1,5,15}, {4,10,2,7,1,5,15}, {5,8,0,1,3,7,15}, {7,10,2,1,3,7,15}, {6,8,0,3,3,7,15}, {4,10,2,3,3,7,15},
+    {5,13,1,1,4,8,0}, {7,15,3,1,4,8,0}, {6,13,1,3,4,8,0}, {4,15,3,3,4,8,0}, {5,13,1,5,6,10,0}, {7,15,3,5,6,10,0},
+    {6,13,1,7,6,10,0}, {4,15,3,7,6,10,0}, {5,12,1,5,5,8,0}, {7,14,3,5,5,8,0}, {6,12,1,7,5,8,0}, {4,14,3,7,5,8,0},
+    {5,12,1,1,7,10,0}, {7,14,3,1,7,10,0}, {6,12,1,3,7,10,0}, {4,14,3,3,7,10,0}, {5,13,5,1,4,12,15}, {7,15,7,1,4,12,15},
+    {6,13,5,3,4,12,15}, {4,15,7,3,4,12,15}, {5,13,5,5,6,14,15}, {7,15,7,5,6,14,15}, {6,13,5,7,6,14,15}, {4,15,7,7,6,14,15},
+    {5,12,5,5,5,12,15}, {7,14,7,5,5,12,15}, {6,12,5,7,5,12,15}, {4,14,7,7,5,12,15}, {5,12,5,1,7,14,15}, {7,14,7,1,7,14,15},
+    {6,12,5,3,7,14,15}, {4,14,7,3,7,14,15}, {5,13,4,1,4,9,0}, {7,15,6,1,4,9,0}, {6,13,4,3,4,9,0}, {4,15,6,3,4,9,0},
+    {5,13,4,5,6,11,0}, {7,15,6,5,6,11,0}, {6,13,4,7,6,11,0}, {4,15,6,7,6,11,0}, {5,12,4,5,5,9,0}, {7,14,6,5,5,9,0},
+    {6,12,4,7,5,9,0}, {4,14,6,7,5,9,0}, {5,12,4,1,7,11,0}, {7,14,6,1,7,11,0}, {6,12,4,3,7,11,0}, {4,14,6,3,7,11,0},
+    {5,13,0,1,4,13,15}, {7,15,2,1,4,13,15}, {6,13,0,3,4,13,15}, {4,15,2,3,4,13,15}, {5,13,0,5,6,15,15}, {7,15,2,5,6,15,15},
+    {6,13,0,7,6,15,15}, {4,15,2,7,6,15,15}, {5,12,0,5,5,13,15}, {7,14,2,5,5,13,15}, {6,12,0,7,5,13,15}, {4,14,2,7,5,13,15},
+    {5,12,0,1,7,15,15}, {7,14,2,1,7,15,15}, {6,12,0,3,7,15,15}, {4,14,2,3,7,15,15}, {5,13,1,0,8,0,0}, {7,15,3,0,8,0,0},
+    {6,13,1,2,8,0,0}, {4,15,3,2,8,0,0}, {5,13,1,4,10,2,0}, {7,15,3,4,10,2,0}, {6,13,1,6,10,2,0}, {4,15,3,6,10,2,0},
+    {5,12,1,4,9,0,0}, {7,14,3,4,9,0,0}, {6,12,1,6,9,0,0}, {4,14,3,6,9,0,0}, {5,12,1,0,11,2,0}, {7,14,3,0,11,2,0},
+    {6,12,1,2,11,2,0}, {4,14,3,2,11,2,0}, {5,13,5,0,8,4,15}, {7,15,7,0,8,4,15}, {6,13,5,2,8,4,15}, {4,15,7,2,8,4,15},
+    {5,13,5,4,10,6,15}, {7,15,7,4,10,6,15}, {6,13,5,6,10,6,15}, {4,15,7,6,10,6,15}, {5,12,5,4,9,4,15}, {7,14,7,4,9,4,15},
+    {6,12,5,6,9,4,15}, {4,14,7,6,9,4,15}, {5,12,5,0,11,6,15}, {7,14,7,0,11,6,15}, {6,12,5,2,11,6,15}, {4,14,7,2,11,6,15},
+    {5,13,4,0,8,1,0}, {7,15,6,0,8,1,0}, {6,13,4,2,8,1,0}, {4,15,6,2,8,1,0}, {5,13,4,4,10,3,0}, {7,15,6,4,10,3,0},
+    {6,13,4,6,10,3,0}, {4,15,6,6,10,3,0}, {5,12,4,4,9,1,0}, {7,14,6,4,9,1,0}, {6,12,4,6,9,1,0}, {4,14,6,6,9,1,0},
+    {5,12,4,0,11,3,0}, {7,14,6,0,11,3,0}, {6,12,4,2,11,3,0}, {4,14,6,2,11,3,0}, {5,13,0,0,8,5,15}, {7,15,2,0,8,5,15},
+    {6,13,0,2,8,5,15}, {4,15,2,2,8,5,15}, {5,13,0,4,10,7,15}, {7,15,2,4,10,7,15}, {6,13,0,6,10,7,15}, {4,15,2,6,10,7,15},
+    {5,12,0,4,9,5,15}, {7,14,2,4,9,5,15}, {6,12,0,6,9,5,15}, {4,14,2,6,9,5,15}, {5,12,0,0,11,7,15}, {7,14,2,0,11,7,15},
+    {6,12,0,2,11,7,15}, {4,14,2,2,11,7,15}, {5,9,1,0,12,8,0}, {7,11,3,0,12,8,0}, {6,9,1,2,12,8,0}, {4,11,3,2,12,8,0},
+    {5,9,1,4,14,10,0}, {7,11,3,4,14,10,0}, {6,9,1,6,14,10,0}, {4,11,3,6,14,10,0}, {5,8,1,4,13,8,0}, {7,10,3,4,13,8,0},
+    {6,8,1,6,13,8,0}, {4,10,3,6,13,8,0}, {5,8,1,0,15,10,0}, {7,10,3,0,15,10,0}, {6,8,1,2,15,10,0}, {4,10,3,2,15,10,0},
+    {5,9,5,0,12,12,15}, {7,11,7,0,12,12,15}, {6,9,5,2,12,12,15}, {4,11,7,2,12,12,15}, {5,9,5,4,14,14,15}, {7,11,7,4,14,14,15},
+    {6,9,5,6,14,14,15}, {4,11,7,6,14,14,15}, {5,8,5,4,13,12,15}, {7,10,7,4,13,12,15}, {6,8,5,6,13,12,15}, {4,10,7,6,13,12,15},
+    {5,8,5,0,15,14,15}, {7,10,7,0,15,14,15}, {6,8,5,2,15,14,15}, {4,10,7,2,15,14,15}, {5,9,4,0,12,9,0}, {7,11,6,0,12,9,0},
+    {6,9,4,2,12,9,0}, {4,11,6,2,12,9,0}, {5,9,4,4,14,11,0}, {7,11,6,4,14,11,0}, {6,9,4,6,14,11,0}, {4,11,6,6,14,11,0},
+    {5,8,4,4,13,9,0}, {7,10,6,4,13,9,0}, {6,8,4,6,13,9,0}, {4,10,6,6,13,9,0}, {5,8,4,0,15,11,0}, {7,10,6,0,15,11,0},
+    {6,8,4,2,15,11,0}, {4,10,6,2,15,11,0}, {5,9,0,0,12,13,15}, {7,11,2,0,12,13,15}, {6,9,0,2,12,13,15}, {4,11,2,2,12,13,15},
+    {5,9,0,4,14,15,15}, {7,11,2,4,14,15,15}, {6,9,0,6,14,15,15}, {4,11,2,6,14,15,15}, {5,8,0,4,13,13,15}, {7,10,2,4,13,13,15},
+    {6,8,0,6,13,13,15}, {4,10,2,6,13,13,15}, {5,8,0,0,15,15,15}, {7,10,2,0,15,15,15}, {6,8,0,2,15,15,15}, {4,10,2,2,15,15,15},
+    {1,9,9,9,0,0,0}, {3,11,11,9,0,0,0}, {2,9,9,11,0,0,0}, {0,11,11,11,0,0,0}, {1,9,9,13,2,2,0}, {3,11,11,13,2,2,0},
+    {2,9,9,15,2,2,0}, {0,11,11,15,2,2,0}, {1,8,9,13,1,0,0}, {3,10,11,13,1,0,0}, {2,8,9,15,1,0,0}, {0,10,11,15,1,0,0},
+    {1,8,9,9,3,2,0}, {3,10,11,9,3,2,0}, {2,8,9,11,3,2,0}, {0,10,11,11,3,2,0}, {1,9,13,9,0,4,15}, {3,11,15,9,0,4,15},
+    {2,9,13,11,0,4,15}, {0,11,15,11,0,4,15}, {1,9,13,13,2,6,15}, {3,11,15,13,2,6,15}, {2,9,13,15,2,6,15}, {0,11,15,15,2,6,15},
+    {1,8,13,13,1,4,15}, {3,10,15,13,1,4,15}, {2,8,13,15,1,4,15}, {0,10,15,15,1,4,15}, {1,8,13,9,3,6,15}, {3,10,15,9,3,6,15},
+    {2,8,13,11,3,6,15}, {0,10,15,11,3,6,15}, {1,9,12,9,0,1,0}, {3,11,14,9,0,1,0}, {2,9,12,11,0,1,0}, {0,11,14,11,0,1,0},
+    {1,9,12,13,2,3,0}, {3,11,14,13,2,3,0}, {2,9,12,15,2,3,0}, {0,11,14,15,2,3,0}, {1,8,12,13,1,1,0}, {3,10,14,13,1,1,0},
+    {2,8,12,15,1,1,0}, {0,10,14,15,1,1,0}, {1,8,12,9,3,3,0}, {3,10,14,9,3,3,0}, {2,8,12,11,3,3,0}, {0,10,14,11,3,3,0},
+    {1,9,8,9,0,5,15}, {3,11,10,9,0,5,15}, {2,9,8,11,0,5,15}, {0,11,10,11,0,5,15}, {1,9,8,13,2,7,15}, {3,11,10,13,2,7,15},
+    {2,9,8,15,2,7,15}, {0,11,10,15,2,7,15}, {1,8,8,13,1,5,15}, {3,10,10,13,1,5,15}, {2,8,8,15,1,5,15}, {0,10,10,15,1,5,15},
+    {1,8,8,9,3,7,15}, {3,10,10,9,3,7,15}, {2,8,8,11,3,7,15}, {0,10,10,11,3,7,15}, {1,13,9,9,4,8,0}, {3,15,11,9,4,8,0},
+    {2,13,9,11,4,8,0}, {0,15,11,11,4,8,0}, {1,13,9,13,6,10,0}, {3,15,11,13,6,10,0}, {2,13,9,15,6,10,0}, {0,15,11,15,6,10,0},
+    {1,12,9,13,5,8,0}, {3,14,11,13,5,8,0}, {2,12,9,15,5,8,0}, {0,14,11,15,5,8,0}, {1,12,9,9,7,10,0}, {3,14,11,9,7,10,0},
+    {2,12,9,11,7,10,0}, {0,14,11,11,7,10,0}, {1,13,13,9,4,12,15}, {3,15,15,9,4,12,15}, {2,13,13,11,4,12,15}, {0,15,15,11,4,12,15},
+    {1,13,13,13,6,14,15}, {3,15,15,13,6,14,15}, {2,13,13,15,6,14,15}, {0,15,15,15,6,14,15}, {1,12,13,13,5,12,15}, {3,14,15,13,5,12,15},
+    {2,12,13,15,5,12,15}, {0,14,15,15,5,12,15}, {1,12,13,9,7,14,15}, {3,14,15,9,7,14,15}, {2,12,13,11,7,14,15}, {0,14,15,11,7,14,15},
+    {1,13,12,9,4,9,0}, {3,15,14,9,4,9,0}, {2,13,12,11,4,9,0}, {0,15,14,11,4,9,0}, {1,13,12,13,6,11,0}, {3,15,14,13,6,11,0},
+    {2,13,12,15,6,11,0}, {0,15,14,15,6,11,0}, {1,12,12,13,5,9,0}, {3,14,14,13,5,9,0}, {2,12,12,15,5,9,0}, {0,14,14,15,5,9,0},
+    {1,12,12,9,7,11,0}, {3,14,14,9,7,11,0}, {2,12,12,11,7,11,0}, {0,14,14,11,7,11,0}, {1,13,8,9,4,13,15}, {3,15,10,9,4,13,15},
+    {2,13,8,11,4,13,15}, {0,15,10,11,4,13,15}, {1,13,8,13,6,15,15}, {3,15,10,13,6,15,15}, {2,13,8,15,6,15,15}, {0,15,10,15,6,15,15},
+    {1,12,8,13,5,13,15}, {3,14,10,13,5,13,15}, {2,12,8,15,5,13,15}, {0,14,10,15,5,13,15}, {1,12,8,9,7,15,15}, {3,14,10,9,7,15,15},
+    {2,12,8,11,7,15,15}, {0,14,10,11,7,15,15}, {1,13,9,8,8,0,0}, {3,15,11,8,8,0,0}, {2,13,9,10,8,0,0}, {0,15,11,10,8,0,0},
+    {1,13,9,12,10,2,0}, {3,15,11,12,10,2,0}, {2,13,9,14,10,2,0}, {0,15,11,14,10,2,0}, {1,12,9,12,9,0,0}, {3,14,11,12,9,0,0},
+    {2,12,9,14,9,0,0}, {0,14,11,14,9,0,0}, {1,12,9,8,11,2,0}, {3,14,11,8,11,2,0}, {2,12,9,10,11,2,0}, {0,14,11,10,11,2,0},
+    {1,13,13,8,8,4,15}, {3,15,15,8,8,4,15}, {2,13,13,10,8,4,15}, {0,15,15,10,8,4,15}, {1,13,13,12,10,6,15}, {3,15,15,12,10,6,15},
+    {2,13,13,14,10,6,15}, {0,15,15,14,10,6,15}, {1,12,13,12,9,4,15}, {3,14,15,12,9,4,15}, {2,12,13,14,9,4,15}, {0,14,15,14,9,4,15},
+    {1,12,13,8,11,6,15}, {3,14,15,8,11,6,15}, {2,12,13,10,11,6,15}, {0,14,15,10,11,6,15}, {1,13,12,8,8,1,0}, {3,15,14,8,8,1,0},
+    {2,13,12,10,8,1,0}, {0,15,14,10,8,1,0}, {1,13,12,12,10,3,0}, {3,15,14,12,10,3,0}, {2,13,12,14,10,3,0}, {0,15,14,14,10,3,0},
+    {1,12,12,12,9,1,0}, {3,14,14,12,9,1,0}, {2,12,12,14,9,1,0}, {0,14,14,14,9,1,0}, {1,12,12,8,11,3,0}, {3,14,14,8,11,3,0},
+    {2,12,12,10,11,3,0}, {0,14,14,10,11,3,0}, {1,13,8,8,8,5,15}, {3,15,10,8,8,5,15}, {2,13,8,10,8,5,15}, {0,15,10,10,8,5,15},
+    {1,13,8,12,10,7,15}, {3,15,10,12,10,7,15}, {2,13,8,14,10,7,15}, {0,15,10,14,10,7,15}, {1,12,8,12,9,5,15}, {3,14,10,12,9,5,15},
+    {2,12,8,14,9,5,15}, {0,14,10,14,9,5,15}, {1,12,8,8,11,7,15}, {3,14,10,8,11,7,15}, {2,12,8,10,11,7,15}, {0,14,10,10,11,7,15},
+    {1,9,9,8,12,8,0}, {3,11,11,8,12,8,0}, {2,9,9,10,12,8,0}, {0,11,11,10,12,8,0}, {1,9,9,12,14,10,0}, {3,11,11,12,14,10,0},
+    {2,9,9,14,14,10,0}, {0,11,11,14,14,10,0}, {1,8,9,12,13,8,0}, {3,10,11,12,13,8,0}, {2,8,9,14,13,8,0}, {0,10,11,14,13,8,0},
+    {1,8,9,8,15,10,0}, {3,10,11,8,15,10,0}, {2,8,9,10,15,10,0}, {0,10,11,10,15,10,0}, {1,9,13,8,12,12,15}, {3,11,15,8,12,12,15},
+    {2,9,13,10,12,12,15}, {0,11,15,10,12,12,15}, {1,9,13,12,14,14,15}, {3,11,15,12,14,14,15}, {2,9,13,14,14,14,15}, {0,11,15,14,14,14,15},
+    {1,8,13,12,13,12,15}, {3,10,15,12,13,12,15}, {2,8,13,14,13,12,15}, {0,10,15,14,13,12,15}, {1,8,13,8,15,14,15}, {3,10,15,8,15,14,15},
+    {2,8,13,10,15,14,15}, {0,10,15,10,15,14,15}, {1,9,12,8,12,9,0}, {3,11,14,8,12,9,0}, {2,9,12,10,12,9,0}, {0,11,14,10,12,9,0},
+    {1,9,12,12,14,11,0}, {3,11,14,12,14,11,0}, {2,9,12,14,14,11,0}, {0,11,14,14,14,11,0}, {1,8,12,12,13,9,0}, {3,10,14,12,13,9,0},
+    {2,8,12,14,13,9,0}, {0,10,14,14,13,9,0}, {1,8,12,8,15,11,0}, {3,10,14,8,15,11,0}, {2,8,12,10,15,11,0}, {0,10,14,10,15,11,0},
+    {1,9,8,8,12,13,15}, {3,11,10,8,12,13,15}, {2,9,8,10,12,13,15}, {0,11,10,10,12,13,15}, {1,9,8,12,14,15,15}, {3,11,10,12,14,15,15},
+    {2,9,8,14,14,15,15}, {0,11,10,14,14,15,15}, {1,8,8,12,13,13,15}, {3,10,10,12,13,13,15}, {2,8,8,14,13,13,15}, {0,10,10,14,13,13,15},
+    {1,8,8,8,15,15,15}, {3,10,10,8,15,15,15}, {2,8,8,10,15,15,15}, {0,10,10,10,15,15,15}
+};
+
+const Dictionary _dict_aruco = Dictionary (&(_dict_aruco_quartets[0][0]), 5, 1024, 1);
+
+
+
+
+}}
+
+#endif // cplusplus
+#endif // __OPENCV_ARUCO_PREDEFINED_DICTIONARIES_CPP__
+
